+++ src/datatrove/pipeline/readers/ipc.py
@@ -13,6 +13,7 @@ class IpcReader(BaseReader):
         self,
         data_folder: BaseInputDataFolder,
         limit: int = -1,
+        stream: bool = False,
         progress: bool = False,
         adapter: Callable = None,
         content_key: str = "content",
@@ -20,20 +21,33 @@ class IpcReader(BaseReader):
         default_metadata: dict = None,
     ):
         super().__init__(data_folder, limit, progress, adapter, content_key, id_key, default_metadata)
+        self.stream = stream
         # TODO: add option to disable reading metadata (https://github.com/apache/arrow/issues/13827 needs to be addressed first)
 
-    def read_file(self, datafile: BaseInputDataFile):
+    @staticmethod
+    def _iter_file_batches(datafile: BaseInputDataFile):
         with datafile.open(binary=True) as f:
             with pa.ipc.open_file(f) as ipc_reader:
-                li = 0
                 for i in range(ipc_reader.num_record_batches):
-                    batch = ipc_reader.get_batch(i)
-                    documents = []
-                    with self.track_time("batch"):
-                        for line in batch.to_pylist():
-                            document = self.get_document_from_dict(line, datafile, li)
-                            if not document:
-                                continue
-                            documents.append(document)
-                            li += 1
-                    yield from documents
+                    yield ipc_reader.get_batch(i)
+
+    @staticmethod
+    def _iter_stream_batches(datafile: BaseInputDataFile):
+        with datafile.open(binary=True) as f:
+            with pa.ipc.open_stream(f) as ipc_stream_reader:
+                for batch in ipc_stream_reader:
+                    yield batch
+
+    def read_file(self, datafile: BaseInputDataFile):
+        batch_iter = self._iter_file_batches(datafile) if not self.stream else self._iter_stream_batches(datafile)
+        li = 0
+        for batch in batch_iter:
+            documents = []
+            with self.track_time("batch"):
+                for line in batch.to_pylist():
+                    document = self.get_document_from_dict(line, datafile, li)
+                    if not document:
+                        continue
+                    documents.append(document)
+                    li += 1
+            yield from documents
